Move fetching access token to separate file and saving it to localStorage
